cmake_modules/sphinx_requirements.txt
alabaster==0.7.16
Babel==2.15.0
beautifulsoup4==4.12.3
brotlicffi==1.1.0.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
#clang==18.1.8
docutils==0.21.2
furo==2024.8.6
idna==3.7
imagesize==1.4.1
Jinja2==3.1.4
libclang==18.1.1
MarkupSafe==2.1.5
packaging==24.1
pycparser==2.22
Pygments==2.18.0
PyStemmer==2.2.0.1
requests==2.32.3
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==8.1.3
sphinx-basic-ng==1.0.0b2
sphinx_inline_tabs==2023.4.21
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sphinxcontrib-websupport==1.2.7
urllib3==2.2.2
